Moved location tests to module in location file
